Refactored "Handlers" into proper folder and fixed imports to reflect
