--- src/seproject/apis/computestore/StorageHandlerAPI.java
@@ -1,5 +1,0 @@
-package seproject.apis.computestore;
-
-public interface StorageHandlerAPI {
-
-}
