+++ src/seproject/apis/usernetworkbridge/handlers/StorageHandlerAPI.java
@@ -1,0 +1,5 @@
+package seproject.apis.usernetworkbridge.handlers;
+
+public interface StorageHandlerAPI {
+
+}
